ItemStackTransformerFeature: Move to ItemStackTransformManager
