+++ common/src/main/java/com/wynntils/features/user/SoulPointTimerFeature.java
@@ -4,11 +4,20 @@
  */
 package com.wynntils.features.user;
 
+import com.google.common.collect.ImmutableList;
 import com.wynntils.core.features.UserFeature;
 import com.wynntils.core.features.properties.FeatureInfo;
 import com.wynntils.core.features.properties.FeatureInfo.Stability;
+import com.wynntils.core.managers.Manager;
+import com.wynntils.wc.custom.item.ItemStackTransformManager;
 
 @FeatureInfo(stability = Stability.STABLE)
 public class SoulPointTimerFeature extends UserFeature {
+    @Override
+    protected void onInit(
+            ImmutableList.Builder<Condition> conditions, ImmutableList.Builder<Class<? extends Manager>> dependencies) {
+        dependencies.add(ItemStackTransformManager.class);
+    }
+
     public static SoulPointTimerFeature INSTANCE;
 }
