+++ common/src/main/java/com/wynntils/features/user/inventory/DurabilityArcFeature.java
@@ -4,14 +4,17 @@
  */
 package com.wynntils.features.user.inventory;
 
+import com.google.common.collect.ImmutableList;
 import com.wynntils.core.config.Config;
 import com.wynntils.core.features.UserFeature;
 import com.wynntils.core.features.properties.FeatureInfo;
 import com.wynntils.core.features.properties.FeatureInfo.Stability;
+import com.wynntils.core.managers.Manager;
 import com.wynntils.mc.event.HotbarSlotRenderEvent;
 import com.wynntils.mc.event.SlotRenderEvent;
 import com.wynntils.mc.render.RenderUtils;
 import com.wynntils.utils.objects.CustomColor;
+import com.wynntils.wc.custom.item.ItemStackTransformManager;
 import com.wynntils.wc.custom.item.WynnItemStack;
 import com.wynntils.wc.custom.item.properties.DurabilityProperty;
 import com.wynntils.wc.custom.item.properties.ItemProperty;
@@ -26,6 +29,12 @@ public class DurabilityArcFeature extends UserFeature {
 
     @Config
     public static boolean renderDurabilityArcHotbar = true;
+
+    @Override
+    protected void onInit(
+            ImmutableList.Builder<Condition> conditions, ImmutableList.Builder<Class<? extends Manager>> dependencies) {
+        dependencies.add(ItemStackTransformManager.class);
+    }
 
     @SubscribeEvent
     public void onRenderHotbarSlot(HotbarSlotRenderEvent.Pre e) {
