+++ common/src/main/java/com/wynntils/features/user/inventory/ItemTextOverlayFeature.java
@@ -4,13 +4,16 @@
  */
 package com.wynntils.features.user.inventory;
 
+import com.google.common.collect.ImmutableList;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.wynntils.core.config.Config;
 import com.wynntils.core.features.UserFeature;
 import com.wynntils.core.features.properties.FeatureInfo;
+import com.wynntils.core.managers.Manager;
 import com.wynntils.mc.event.HotbarSlotRenderEvent;
 import com.wynntils.mc.event.SlotRenderEvent;
 import com.wynntils.mc.render.FontRenderer;
+import com.wynntils.wc.custom.item.ItemStackTransformManager;
 import com.wynntils.wc.custom.item.WynnItemStack;
 import com.wynntils.wc.custom.item.properties.ItemProperty;
 import com.wynntils.wc.custom.item.properties.type.TextOverlayProperty;
@@ -60,6 +63,12 @@ public class ItemTextOverlayFeature extends UserFeature {
 
     @Config
     public static boolean hotbarTextOverlayEnabled = true;
+
+    @Override
+    protected void onInit(
+            ImmutableList.Builder<Condition> conditions, ImmutableList.Builder<Class<? extends Manager>> dependencies) {
+        dependencies.add(ItemStackTransformManager.class);
+    }
 
     @SubscribeEvent
     public void onRenderSlot(SlotRenderEvent.Post e) {
